code/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio

numpy
pillow
matplotlib
colorama
rich
Levenshtein
tqdm
seaborn
scikit-learn
